requirements.txt
streamlit==1.33.0
pandas>=2.0.0
altair==5.3.0
dbt-duckdb==1.9.0
watchdog==4.0.0
matplotlib>=3.8.0
